+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- Fix Mathjax equations not displayed properly (#283)
+
 ## [2.1.2] - 2024-05-13
 
 ## Fixed